// src/app/models/place-order.ts
export default class PlaceOrder{

    OrderId:string="";
    CartItemId:string="";
    RestaurantId:string ="";
    RestaurantName: string="";
    MenuId: string="";
    MenuName: string="";
    ProductId: string="";
    ProductName: string="";
    Price: number=0;
    ItemCount: number=0;
    Amount: number=0;
    UserId: string="";
    UserName: string="";
    MobileNo: string="";
    Address: string="";
    CreatedDate: String="";
    CreatedBy: string="";
    Status: string="";
    ActiveYn:boolean=true;
    DeleteYn:boolean=false;
    
 }